+++ R/shinyjs.R
@@ -5,9 +5,7 @@
 #' \code{shinyjs} lets you perform common JavaScript operations in Shiny
 #' applications without having to know any JavaScript. Examples include hiding an
 #' element, disabling an input, resetting an input back to its original value,
-#' and many more useful functions. Even if you do know JavaScript,
-#' \code{shinyjs} can be used as convenience functions to avoid dealing with message
-#' passing and writing JavaScript code. \code{shinyjs} can also be used to
+#' and many more useful functions. \code{shinyjs} can also be used to
 #' easily run your own custom JavaScript functions from R.
 #'
 #' View a \href{http://daattali.com/shiny/shinyjs-demo/}{demo Shiny app}